requirements.txt
numpy==1.26.4
torch==2.2.2
torchvision
torchaudio
einops
huggingface-hub>=0.22.0
imageio
matplotlib
ninja
opencv-python
pandas
safetensors
scipy
timm
tqdm
wandb
xformers==0.0.25.post1
neptune
scikit-image
plyfile
torchmetrics
omegaconf
hydra-core
jaxtyping
hydra-submitit-launcher
submitit
diff-gaussian-rasterization @ git+https://github.com/eldar/diff-gaussian-rasterization-w-pose@main
GitPython
lightning
ema-pytorch
pytubefix
